% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.StepGraph.R
\name{cv.stepGraph}
\alias{cv.stepGraph}
\title{Cross-Validation for Stepwise Gaussian Graphical Model}
\usage{
cv.stepGraph(x, fold, alpha_f_min, alpha_f_max, n_alpha, nei.max)
}
\arguments{
\item{x}{Data matrix (of size n x p).}

\item{fold}{Number of folds for the cross-validation procedure.}

\item{alpha_f_min}{Minimum threshold value for the cross-validation procedure.}

\item{alpha_f_max}{Minimum threshold value for the cross-validation procedure.}

\item{n_alpha}{Number of elements in the grid for the cross-validation.}

\item{nei.max}{Maximum number of variables in every neighborhood.}
}
\value{
A list
}
\description{
\code{cv.stepGraph} implements the cross-valiation procedure for the stepwise gaussian graphical algorithm.
}
\author{
Anthony-Alexander Christidis, \email{anthony.christidis@stat.ubc.ca}
}
